+++ libraries/stdlib/samples/test/samples/collections/iterables.kt
@@ -17,7 +17,7 @@
 package samples.collections
 
 import samples.*
-import kotlin.coroutines.experimental.buildIterator
+import kotlin.coroutines.experimental.defineIterator
 
 @RunWith(Enclosed::class)
 class Iterables {
@@ -27,7 +27,7 @@ class Iterables {
         @Sample
         fun iterable() {
             val iterable = Iterable {
-                buildIterator {
+                defineIterator {
                     yield(42)
                     yieldAll(1..5 step 2)
                 }