+++ libraries/stdlib/src/kotlin/collections/SlidingWindow.kt
@@ -6,7 +6,7 @@
 package kotlin.collections
 
 import kotlin.*
-import kotlin.coroutines.experimental.buildIterator
+import kotlin.coroutines.experimental.defineIterator
 
 internal fun checkWindowSizeStep(size: Int, step: Int) {
     require(size > 0 && step > 0) {
@@ -24,7 +24,7 @@ internal fun <T> Sequence<T>.windowedSequence(size: Int, step: Int, partialWindo
 
 internal fun <T> windowedIterator(iterator: Iterator<T>, size: Int, step: Int, partialWindows: Boolean, reuseBuffer: Boolean): Iterator<List<T>> {
     if (!iterator.hasNext()) return EmptyIterator
-    return buildIterator<List<T>> {
+    return defineIterator<List<T>> {
         val gap = step - size
         if (gap >= 0) {
             var buffer = ArrayList<T>(size)